+++ src/zope/testing/formparser.py
@@ -8,17 +8,22 @@ See :doc:`../formparser` for documentation.
 This isn't intended to simulate a browser session; that's provided by
 the `zope.testbrowser` package.
 
+.. versionchanged:: 4.10.0
+   Add support for Python 3.
+
 """
 __docformat__ = "reStructuredText"
 
-import HTMLParser
-import urlparse
-import warnings
 
+try:
+    import html.parser as HTMLParser
+except ImportError:  # Python 2
+    import HTMLParser
 
-warnings.warn(
-    'zope.testing.formparser is deprecated. It does not work on Python 3.',
-    DeprecationWarning, stacklevel=2)
+try:
+    import urllib.parse as urlparse
+except ImportError:  # Python 2
+    import urlparse
 
 
 def parse(data, base=None):